mod: Payment Service 인터페이스 추상화 및 에러 핸들링 코드 추가
